remove old join impl
