+++ crates/polars-stream/src/nodes/joins/equi_join.rs
@@ -1,22 +1,25 @@
+use std::cmp::Reverse;
+use std::collections::BinaryHeap;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::{Arc, LazyLock};
 
+use arrow::array::builder::ShareStrategy;
 use crossbeam_queue::ArrayQueue;
+use polars_core::frame::builder::DataFrameBuilder;
 use polars_core::prelude::*;
 use polars_core::schema::{Schema, SchemaExt};
-use polars_core::series::IsSorted;
-use polars_core::utils::accumulate_dataframes_vertical_unchecked;
 use polars_core::{config, POOL};
-use polars_expr::chunked_idx_table::{new_chunked_idx_table, ChunkedIdxTable};
 use polars_expr::hash_keys::HashKeys;
+use polars_expr::idx_table::{new_idx_table, IdxTable};
 use polars_io::pl_async::get_runtime;
 use polars_ops::frame::{JoinArgs, JoinType, MaintainOrderJoin};
-use polars_ops::prelude::TakeChunked;
 use polars_ops::series::coalesce_columns;
 use polars_utils::cardinality_sketch::CardinalitySketch;
 use polars_utils::hashing::HashPartitioner;
 use polars_utils::itertools::Itertools;
 use polars_utils::pl_str::PlSmallStr;
+use polars_utils::priority::Priority;
+use polars_utils::sparse_init_vec::SparseInitVec;
 use polars_utils::{format_pl_smallstr, IdxSize};
 use rayon::prelude::*;
 
@@ -311,7 +314,7 @@ impl SampleState {
         recv: &[PortState],
         num_pipelines: usize,
         params: &mut EquiJoinParams,
-        table: &mut Option<Box<dyn ChunkedIdxTable>>,
+        table: &mut Option<Box<dyn IdxTable>>,
     ) -> PolarsResult<Option<BuildState>> {
         let left_saturated = self.left_len >= *SAMPLE_LIMIT;
         let right_saturated = self.right_len >= *SAMPLE_LIMIT;
@@ -401,9 +404,9 @@ impl SampleState {
         // Transition to building state.
         params.left_is_build = Some(left_is_build);
         *table = Some(if left_is_build {
-            new_chunked_idx_table(params.left_key_schema.clone())
+            new_idx_table(params.left_key_schema.clone())
         } else {
-            new_chunked_idx_table(params.right_key_schema.clone())
+            new_idx_table(params.right_key_schema.clone())
         });
 
         let mut sampled_build_morsels =
@@ -416,7 +419,9 @@ impl SampleState {
 
         let partitioner = HashPartitioner::new(num_pipelines, 0);
         let mut build_state = BuildState {
-            partitions_per_worker: (0..num_pipelines).map(|_| Vec::new()).collect(),
+            local_builders: (0..num_pipelines)
+                .map(|_| LocalBuilder::default())
+                .collect(),
             sampled_probe_morsels,
         };
 
@@ -429,13 +434,12 @@ impl SampleState {
                     .reinsert(num_pipelines, None, scope, &mut join_handles)
                     .unwrap();
 
-                for (worker_ps, recv) in build_state.partitions_per_worker.iter_mut().zip(receivers)
-                {
+                for (local_builder, recv) in build_state.local_builders.iter_mut().zip(receivers) {
                     join_handles.push(scope.spawn_task(
                         TaskPriority::High,
                         BuildState::partition_and_sink(
                             recv,
-                            worker_ps,
+                            local_builder,
                             partitioner.clone(),
                             params,
                             &state,
@@ -457,30 +461,48 @@ impl SampleState {
 }
 
 #[derive(Default)]
-struct BuildPartition {
-    hash_keys: Vec<HashKeys>,
-    frames: Vec<(MorselSeq, DataFrame)>,
-    sketch: Option<CardinalitySketch>,
+struct LocalBuilder {
+    // The complete list of morsels and their computed hashes seen by this builder.
+    morsels: Vec<(MorselSeq, DataFrame, HashKeys)>,
+
+    // A cardinality sketch per partition for the keys seen by this builder.
+    sketch_per_p: Vec<CardinalitySketch>,
+
+    // morsel_idxs_values_per_p[p][start..stop] contains the offsets into morsels[i]
+    // for partition p, where start, stop are:
+    // let start = morsel_idxs_offsets[i * num_partitions + p];
+    // let stop = morsel_idxs_offsets[(i + 1) * num_partitions + p];
+    morsel_idxs_values_per_p: Vec<Vec<IdxSize>>,
+    morsel_idxs_offsets_per_p: Vec<usize>,
 }
 
 #[derive(Default)]
 struct BuildState {
-    partitions_per_worker: Vec<Vec<BuildPartition>>,
+    local_builders: Vec<LocalBuilder>,
     sampled_probe_morsels: BufferedStream,
 }
 
 impl BuildState {
     async fn partition_and_sink(
         mut recv: Receiver<Morsel>,
-        partitions: &mut Vec<BuildPartition>,
+        local: &mut LocalBuilder,
         partitioner: HashPartitioner,
         params: &EquiJoinParams,
         state: &ExecutionState,
     ) -> PolarsResult<()> {
         let track_unmatchable = params.emit_unmatched_build();
-        let mut partition_idxs = vec![Vec::new(); partitioner.num_partitions()];
-        partitions.resize_with(partitioner.num_partitions(), BuildPartition::default);
-        let mut sketches = vec![CardinalitySketch::default(); partitioner.num_partitions()];
+        local
+            .sketch_per_p
+            .resize_with(partitioner.num_partitions(), Default::default);
+        local
+            .morsel_idxs_values_per_p
+            .resize_with(partitioner.num_partitions(), Default::default);
+
+        if local.morsel_idxs_offsets_per_p.is_empty() {
+            local
+                .morsel_idxs_offsets_per_p
+                .resize(partitioner.num_partitions(), 0);
+        }
 
         let (key_selectors, payload_selector);
         if params.left_is_build.unwrap() {
@@ -493,140 +515,243 @@ impl BuildState {
 
         while let Ok(morsel) = recv.recv().await {
             // Compute hashed keys and payload. We must rechunk the payload for
-            // later chunked gathers.
+            // later gathers.
             let hash_keys = select_keys(morsel.df(), key_selectors, params, state).await?;
             let mut payload = select_payload(morsel.df().clone(), payload_selector);
             payload.rechunk_mut();
-            payload._deshare_views_mut();
 
-            unsafe {
-                for p in partition_idxs.iter_mut() {
-                    p.clear();
-                }
-                hash_keys.gen_idxs_per_partition(
-                    &partitioner,
-                    &mut partition_idxs,
-                    &mut sketches,
-                    track_unmatchable,
-                );
-                for (p, idxs_in_p) in partitions.iter_mut().zip(&partition_idxs) {
-                    let payload_for_partition = payload.take_slice_unchecked_impl(idxs_in_p, false);
-                    p.hash_keys.push(hash_keys.gather(idxs_in_p));
-                    p.frames.push((morsel.seq(), payload_for_partition));
-                }
-            }
+            hash_keys.gen_idxs_per_partition(
+                &partitioner,
+                &mut local.morsel_idxs_values_per_p,
+                &mut local.sketch_per_p,
+                track_unmatchable,
+            );
+
+            local
+                .morsel_idxs_offsets_per_p
+                .extend(local.morsel_idxs_values_per_p.iter().map(|vp| vp.len()));
+            local.morsels.push((morsel.seq(), payload, hash_keys));
         }
-
-        for (p, sketch) in sketches.into_iter().enumerate() {
-            partitions[p].sketch = Some(sketch);
-        }
-
         Ok(())
     }
 
-    fn finalize(&mut self, params: &EquiJoinParams, table: &dyn ChunkedIdxTable) -> ProbeState {
-        // Transpose.
-        let num_workers = self.partitions_per_worker.len();
-        let num_partitions = self.partitions_per_worker[0].len();
-        let mut results_per_partition = (0..num_partitions)
-            .map(|_| Vec::with_capacity(num_workers))
-            .collect_vec();
-        for worker in self.partitions_per_worker.drain(..) {
-            for (p, result) in worker.into_iter().enumerate() {
-                results_per_partition[p].push(result);
-            }
-        }
+    fn finalize_ordered(&mut self, params: &EquiJoinParams, table: &dyn IdxTable) -> ProbeState {
+        let track_unmatchable = params.emit_unmatched_build();
+        let payload_schema = if params.left_is_build.unwrap() {
+            &params.left_payload_schema
+        } else {
+            &params.right_payload_schema
+        };
 
-        POOL.install(|| {
-            let track_unmatchable = params.emit_unmatched_build();
-            let table_per_partition: Vec<_> = results_per_partition
-                .into_par_iter()
-                .with_max_len(1)
-                .map(|results| {
-                    // Estimate sizes and cardinality.
+        let num_partitions = self.local_builders[0].sketch_per_p.len();
+        let local_builders = &self.local_builders;
+        let probe_tables: SparseInitVec<ProbeTable> = SparseInitVec::with_capacity(num_partitions);
+        
+        POOL.scope(|s| {
+            for p in 0..num_partitions {
+                let probe_tables = &probe_tables;
+                s.spawn(move |_| {
+                    // TODO: every thread does an identical linearize, we can do a single parallel one.
+                    let mut kmerge = BinaryHeap::with_capacity(local_builders.len());
+                    let mut cur_idx_per_loc = vec![0; local_builders.len()];
+
+                    // Compute cardinality estimate and total amount of
+                    // payload for this partition, and initialize k-way merge.
                     let mut sketch = CardinalitySketch::new();
-                    let mut num_frames = 0;
-                    for result in &results {
-                        sketch.combine(result.sketch.as_ref().unwrap());
-                        num_frames += result.frames.len();
+                    let mut payload_rows = 0;
+                    for (l_idx, l) in local_builders.iter().enumerate() {
+                        let Some((seq, _, _)) = l.morsels.get(0) else { continue };
+                        kmerge.push(Priority(Reverse(seq), l_idx));
+
+                        sketch.combine(&l.sketch_per_p[p]);
+                        let offsets_len = l.morsel_idxs_offsets_per_p.len();
+                        payload_rows +=
+                            l.morsel_idxs_offsets_per_p[offsets_len - num_partitions + p];
                     }
 
-                    // Build table for this partition.
-                    let mut combined_frames = Vec::with_capacity(num_frames);
-                    let mut chunk_seq_ids = Vec::with_capacity(num_frames);
-                    let mut table = table.new_empty();
-                    table.reserve(sketch.estimate() * 5 / 4);
-                    if params.preserve_order_build {
-                        let mut combined = Vec::with_capacity(num_frames);
-                        for result in results {
-                            for (hash_keys, (seq, frame)) in
-                                result.hash_keys.into_iter().zip(result.frames)
-                            {
-                                combined.push((seq, hash_keys, frame));
+                    // Allocate hash table and payload builder.
+                    let mut p_table = table.new_empty();
+                    p_table.reserve(sketch.estimate() * 5 / 4);
+                    let mut p_payload = DataFrameBuilder::new(payload_schema.clone());
+                    p_payload.reserve(payload_rows);
+                    
+                    let mut p_seq_ids = Vec::new();
+                    if track_unmatchable {
+                        p_seq_ids.reserve(payload_rows);
+                    }
+
+                    // Linearize and build.
+                    unsafe {
+                        let mut norm_seq_id = 0 as IdxSize;
+                        while let Some(Priority(Reverse(seq), l_idx)) = kmerge.pop() {
+                            let l = local_builders.get_unchecked(l_idx);
+                            let idx_in_l = *cur_idx_per_loc.get_unchecked(l_idx);
+                            *cur_idx_per_loc.get_unchecked_mut(l_idx) += 1;
+                            if let Some((next_seq, _, _)) = l.morsels.get(idx_in_l + 1) {
+                                kmerge.push(Priority(Reverse(next_seq), l_idx));
                             }
-                        }
-
-                        combined.sort_unstable_by_key(|c| c.0);
-                        for (seq, hash_keys, frame) in combined {
-                            // Zero-sized chunks can get deleted, so skip entirely to avoid messing
-                            // up the chunk counter.
-                            if frame.height() == 0 {
-                                continue;
-                            }
-
-                            table.insert_key_chunk(hash_keys, track_unmatchable);
-                            combined_frames.push(frame);
-                            chunk_seq_ids.push(seq);
-                        }
-                    } else {
-                        for result in results {
-                            for (hash_keys, (_, frame)) in
-                                result.hash_keys.into_iter().zip(result.frames)
-                            {
-                                // Zero-sized chunks can get deleted, so skip entirely to avoid messing
-                                // up the chunk counter.
-                                if frame.height() == 0 {
-                                    continue;
-                                }
-
-                                table.insert_key_chunk(hash_keys, track_unmatchable);
-                                combined_frames.push(frame);
+                            
+                            let (_mseq, payload, keys) = l.morsels.get_unchecked(idx_in_l);
+                            let p_morsel_idxs_start =
+                                l.morsel_idxs_offsets_per_p[idx_in_l * num_partitions + p];
+                            let p_morsel_idxs_stop =
+                                l.morsel_idxs_offsets_per_p[(idx_in_l + 1) * num_partitions + p];
+                            let p_morsel_idxs = &l.morsel_idxs_values_per_p[p]
+                                [p_morsel_idxs_start..p_morsel_idxs_stop];
+                            p_table.insert_keys_subset(keys, p_morsel_idxs, track_unmatchable);
+                            p_payload.gather_extend(
+                                payload,
+                                p_morsel_idxs,
+                                ShareStrategy::Never,
+                            );
+                            
+                            if track_unmatchable {
+                                p_seq_ids.resize(p_payload.len(), norm_seq_id);
+                                norm_seq_id += 1;
                             }
                         }
                     }
 
-                    let df = if combined_frames.is_empty() {
-                        if params.left_is_build.unwrap() {
-                            DataFrame::empty_with_schema(&params.left_payload_schema)
-                        } else {
-                            DataFrame::empty_with_schema(&params.right_payload_schema)
-                        }
-                    } else {
-                        accumulate_dataframes_vertical_unchecked(combined_frames)
-                    };
-                    ProbeTable {
-                        hash_table: table,
-                        payload: df,
-                        chunk_seq_ids,
-                    }
-                })
-                .collect();
-
-            ProbeState {
-                table_per_partition,
-                max_seq_sent: MorselSeq::default(),
-                sampled_probe_morsels: core::mem::take(&mut self.sampled_probe_morsels),
+                    probe_tables.try_set(p, ProbeTable {
+                        hash_table: p_table,
+                        payload: p_payload.freeze(),
+                        seq_ids: p_seq_ids,
+                    }).ok().unwrap();
+                });
             }
-        })
+        });
+
+        ProbeState {
+            table_per_partition: probe_tables.try_assume_init().ok().unwrap(),
+            max_seq_sent: MorselSeq::default(),
+            sampled_probe_morsels: core::mem::take(&mut self.sampled_probe_morsels),
+        }
+    }
+
+    fn finalize_unordered(&mut self, params: &EquiJoinParams, table: &dyn IdxTable) -> ProbeState {
+        let track_unmatchable = params.emit_unmatched_build();
+        let payload_schema = if params.left_is_build.unwrap() {
+            &params.left_payload_schema
+        } else {
+            &params.right_payload_schema
+        };
+
+        // To reduce maximum memory usage we want to drop the morsels
+        // as soon as they're processed, so we move into Arcs. The drops might
+        // also be expensive, so instead of directly dropping we put that on
+        // a work queue.
+        let morsels_per_local_builder = self
+            .local_builders
+            .iter_mut()
+            .map(|b| Arc::new(core::mem::take(&mut b.morsels)))
+            .collect_vec();
+        let (morsel_drop_q_send, morsel_drop_q_recv) = crossbeam_channel::bounded(morsels_per_local_builder.len());
+        let num_partitions = self.local_builders[0].sketch_per_p.len();
+        let local_builders = &self.local_builders;
+        let probe_tables: SparseInitVec<ProbeTable> = SparseInitVec::with_capacity(num_partitions);
+
+        POOL.scope(|s| {
+            // Wrap in outer Arc to move to each thread, performing the
+            // expensive clone on that thread.
+            let arc_morsels_per_local_builder = Arc::new(morsels_per_local_builder);
+            for p in 0..num_partitions {
+                let arc_morsels_per_local_builder = Arc::clone(&arc_morsels_per_local_builder);
+                let morsel_drop_q_send = morsel_drop_q_send.clone();
+                let morsel_drop_q_recv = morsel_drop_q_recv.clone();
+                let probe_tables = &probe_tables;
+                s.spawn(move |_| {
+                    // Extract from outer arc and drop outer arc.
+                    let morsels_per_local_builder =
+                        Arc::unwrap_or_clone(arc_morsels_per_local_builder);
+
+                    // Compute cardinality estimate and total amount of
+                    // payload for this partition.
+                    let mut sketch = CardinalitySketch::new();
+                    let mut payload_rows = 0;
+                    for l in local_builders {
+                        sketch.combine(&l.sketch_per_p[p]);
+                        let offsets_len = l.morsel_idxs_offsets_per_p.len();
+                        payload_rows +=
+                            l.morsel_idxs_offsets_per_p[offsets_len - num_partitions + p];
+                    }
+
+                    // Allocate hash table and payload builder.
+                    let mut p_table = table.new_empty();
+                    p_table.reserve(sketch.estimate() * 5 / 4);
+                    let mut p_payload = DataFrameBuilder::new(payload_schema.clone());
+                    p_payload.reserve(payload_rows);
+
+                    // Build.
+                    let mut skip_drop_attempt = false;
+                    for (l, l_morsels) in local_builders.iter().zip(morsels_per_local_builder) {
+                        // Try to help with dropping the processed morsels.
+                        if !skip_drop_attempt {
+                            drop(morsel_drop_q_recv.try_recv());
+                        }
+
+                        for (i, morsel) in l_morsels.iter().enumerate() {
+                            let (_mseq, payload, keys) = morsel;
+                            unsafe {
+                                let p_morsel_idxs_start =
+                                    l.morsel_idxs_offsets_per_p[i * num_partitions + p];
+                                let p_morsel_idxs_stop =
+                                    l.morsel_idxs_offsets_per_p[(i + 1) * num_partitions + p];
+                                let p_morsel_idxs = &l.morsel_idxs_values_per_p[p]
+                                    [p_morsel_idxs_start..p_morsel_idxs_stop];
+                                p_table.insert_keys_subset(keys, p_morsel_idxs, track_unmatchable);
+                                p_payload.gather_extend(
+                                    payload,
+                                    p_morsel_idxs,
+                                    ShareStrategy::Never,
+                                );
+                            }
+                        }
+                        
+                        if let Some(l) = Arc::into_inner(l_morsels) {
+                            // If we're the last thread to process this set of morsels we're probably
+                            // falling behind the rest, since the drop can be quite expensive we skip
+                            // a drop attempt hoping someone else will pick up the slack.
+                            morsel_drop_q_send.send(l).unwrap();
+                            skip_drop_attempt = true;
+                        } else {
+                            skip_drop_attempt = false;
+                        }
+                    }
+                    
+                    // We're done, help others out by doing drops.
+                    drop(morsel_drop_q_send); // So we don't deadlock trying to receive from ourselves.
+                    while let Ok(l_morsels) = morsel_drop_q_recv.recv() {
+                        drop(l_morsels);
+                    }
+
+                    probe_tables.try_set(p, ProbeTable {
+                        hash_table: p_table,
+                        payload: p_payload.freeze(),
+                        seq_ids: Vec::new(),
+                    }).ok().unwrap();
+                });
+            }
+
+            // Drop outer arc after spawning each thread so the inner arcs
+            // can get dropped as soon as they're processed. We also have to
+            // drop the drop queue sender so we don't deadlock waiting for it
+            // to end.
+            drop(arc_morsels_per_local_builder);
+            drop(morsel_drop_q_send);
+        });
+
+        ProbeState {
+            table_per_partition: probe_tables.try_assume_init().ok().unwrap(),
+            max_seq_sent: MorselSeq::default(),
+            sampled_probe_morsels: core::mem::take(&mut self.sampled_probe_morsels),
+        }
     }
 }
 
 struct ProbeTable {
-    // Important that df is not rechunked, the chunks it was inserted with
-    // into the table must be preserved for chunked gathers.
-    hash_table: Box<dyn ChunkedIdxTable>,
+    hash_table: Box<dyn IdxTable>,
     payload: DataFrame,
-    chunk_seq_ids: Vec<MorselSeq>,
+    seq_ids: Vec<IdxSize>,
 }
 
 struct ProbeState {
@@ -647,6 +772,8 @@ impl ProbeState {
     ) -> PolarsResult<MorselSeq> {
         // TODO: shuffle after partitioning and keep probe tables thread-local.
         let mut partition_idxs = vec![Vec::new(); partitioner.num_partitions()];
+        let mut probe_partitions = Vec::new();
+        let mut materialized_idxsize_range = Vec::new();
         let mut table_match = Vec::new();
         let mut probe_match = Vec::new();
         let mut max_seq = MorselSeq::default();
@@ -655,115 +782,152 @@ impl ProbeState {
         let mark_matches = params.emit_unmatched_build();
         let emit_unmatched = params.emit_unmatched_probe();
 
-        let (key_selectors, payload_selector);
+        let (key_selectors, payload_selector, build_payload_schema, probe_payload_schema);
         if params.left_is_build.unwrap() {
-            payload_selector = &params.right_payload_select;
             key_selectors = &params.right_key_selectors;
+            payload_selector = &params.right_payload_select;
+            build_payload_schema = &params.left_payload_schema;
+            probe_payload_schema = &params.right_payload_schema;
         } else {
-            payload_selector = &params.left_payload_select;
             key_selectors = &params.left_key_selectors;
+            payload_selector = &params.left_payload_select;
+            build_payload_schema = &params.right_payload_schema;
+            probe_payload_schema = &params.left_payload_schema;
         };
+
+        let mut build_out = DataFrameBuilder::new(build_payload_schema.clone());
+        let mut probe_out = DataFrameBuilder::new(probe_payload_schema.clone());
+
+        // A simple estimate used to size reserves.
+        let mut selectivity_estimate = 1.0;
 
         while let Ok(morsel) = recv.recv().await {
             // Compute hashed keys and payload.
             let (df, seq, src_token, wait_token) = morsel.into_inner();
+            max_seq = seq;
+
+            let df_height = df.height();
+            if df_height == 0 {
+                continue;
+            }
+
             let hash_keys = select_keys(&df, key_selectors, params, state).await?;
             let mut payload = select_payload(df, payload_selector);
             let mut payload_rechunked = false; // We don't eagerly rechunk because there might be no matches.
+            let mut total_matches = 0;
 
-            max_seq = seq;
+            // Use selectivity estimate to reserve for morsel builders.
+            let max_match_per_key_est = selectivity_estimate as usize + 16;
+            let out_est_size = ((selectivity_estimate * 1.2 * df_height as f64) as usize)
+                .min(probe_limit as usize);
+            build_out.reserve(out_est_size + max_match_per_key_est);
 
             unsafe {
-                // Partition and probe the tables.
-                for p in partition_idxs.iter_mut() {
-                    p.clear();
-                }
-                hash_keys.gen_idxs_per_partition(
-                    &partitioner,
-                    &mut partition_idxs,
-                    &mut [],
-                    emit_unmatched,
-                );
+                let new_morsel = |build: &mut DataFrameBuilder, probe: &mut DataFrameBuilder| {
+                    let mut build_df = build.freeze_reset();
+                    let mut probe_df = probe.freeze_reset();
+                    let out_df = if params.left_is_build.unwrap() {
+                        build_df.hstack_mut_unchecked(probe_df.get_columns());
+                        build_df
+                    } else {
+                        probe_df.hstack_mut_unchecked(build_df.get_columns());
+                        probe_df
+                    };
+                    let out_df = postprocess_join(out_df, params);
+                    Morsel::new(out_df, seq, src_token.clone())
+                };
+
                 if params.preserve_order_probe {
-                    // TODO: non-sort based implementation, can directly scatter
-                    // after finding matches for each partition.
-                    let mut out_per_partition = Vec::with_capacity(partitioner.num_partitions());
-                    let name = PlSmallStr::from_static("__POLARS_PROBE_PRESERVE_ORDER_IDX");
-                    for (p, idxs_in_p) in partitions.iter().zip(&partition_idxs) {
-                        table_match.clear();
-                        probe_match.clear();
-                        p.hash_table.probe_subset(
-                            &hash_keys,
-                            idxs_in_p,
-                            &mut table_match,
-                            &mut probe_match,
-                            mark_matches,
-                            emit_unmatched,
-                            IdxSize::MAX,
+                    // To preserve the order we can't do bulk probes per partition and must follow
+                    // the order of the probe morsel. We can still group probes that are
+                    // consecutively on the same partition.
+                    probe_partitions.clear();
+                    hash_keys.gen_partitions(&partitioner, &mut probe_partitions, emit_unmatched);
+                    let mut probe_group_start = 0;
+                    while probe_group_start < probe_partitions.len() {
+                        let p_idx = probe_partitions[probe_group_start];
+                        let mut probe_group_end = probe_group_start + 1;
+                        while probe_partitions.get(probe_group_end) == Some(&p_idx) {
+                            probe_group_end += 1;
+                        }
+                        let Some(p) = partitions.get(p_idx as usize) else {
+                            probe_group_start = probe_group_end;
+                            continue;
+                        };
+
+                        materialized_idxsize_range.extend(
+                            materialized_idxsize_range.len() as IdxSize..probe_group_end as IdxSize,
                         );
 
-                        if table_match.is_empty() {
-                            continue;
-                        }
+                        while probe_group_start < probe_group_end {
+                            let matches_before_limit = probe_limit - probe_match.len() as IdxSize;
+                            table_match.clear();
+                            probe_group_start += p.hash_table.probe_subset(
+                                &hash_keys,
+                                &materialized_idxsize_range[probe_group_start..probe_group_end],
+                                &mut table_match,
+                                &mut probe_match,
+                                mark_matches,
+                                emit_unmatched,
+                                matches_before_limit,
+                            ) as usize;
 
-                        // Gather output and add to buffer.
-                        let mut build_df = if emit_unmatched {
-                            p.payload.take_opt_chunked_unchecked(&table_match, false)
-                        } else {
-                            p.payload
-                                .take_chunked_unchecked(&table_match, IsSorted::Not, false)
-                        };
+                            if emit_unmatched {
+                                build_out.opt_gather_extend(
+                                    &p.payload,
+                                    &table_match,
+                                    ShareStrategy::Always,
+                                );
+                            } else {
+                                build_out.gather_extend(
+                                    &p.payload,
+                                    &table_match,
+                                    ShareStrategy::Always,
+                                );
+                            };
 
-                        if !payload_rechunked {
-                            // TODO: can avoid rechunk? We have to rechunk here or else we do it
-                            // multiple times during the gather.
-                            payload.rechunk_mut();
-                            payload_rechunked = true;
-                        }
-                        let mut probe_df = payload.take_slice_unchecked_impl(&probe_match, false);
-
-                        let mut out_df = if params.left_is_build.unwrap() {
-                            build_df.hstack_mut_unchecked(probe_df.get_columns());
-                            build_df
-                        } else {
-                            probe_df.hstack_mut_unchecked(build_df.get_columns());
-                            probe_df
-                        };
-
-                        let idxs_ca =
-                            IdxCa::from_vec(name.clone(), core::mem::take(&mut probe_match));
-                        out_df.with_column_unchecked(idxs_ca.into_column());
-                        out_per_partition.push(out_df);
-                    }
-
-                    if !out_per_partition.is_empty() {
-                        let sort_options = SortMultipleOptions {
-                            descending: vec![false],
-                            nulls_last: vec![false],
-                            multithreaded: false,
-                            maintain_order: true,
-                            limit: None,
-                        };
-                        let mut out_df =
-                            accumulate_dataframes_vertical_unchecked(out_per_partition);
-                        out_df.sort_in_place([name.clone()], sort_options).unwrap();
-                        out_df.drop_in_place(&name).unwrap();
-                        out_df = postprocess_join(out_df, params);
-
-                        // TODO: break in smaller morsels.
-                        let out_morsel = Morsel::new(out_df, seq, src_token.clone());
-                        if send.send(out_morsel).await.is_err() {
-                            break;
+                            if probe_match.len() >= probe_limit as usize
+                                || probe_group_start == probe_partitions.len()
+                            {
+                                if !payload_rechunked {
+                                    payload.rechunk_mut();
+                                    payload_rechunked = true;
+                                }
+                                probe_out.gather_extend(
+                                    &payload,
+                                    &probe_match,
+                                    ShareStrategy::Always,
+                                );
+                                probe_match.clear();
+                                let out_morsel = new_morsel(&mut build_out, &mut probe_out);
+                                if send.send(out_morsel).await.is_err() {
+                                    return Ok(max_seq);
+                                }
+                                if probe_group_end != probe_partitions.len() {
+                                    // We had enough matches to need a mid-partition flush, let's assume there are a lot of
+                                    // matches and just do a large reserve.
+                                    build_out.reserve(probe_limit as usize + max_match_per_key_est);
+                                }
+                            }
                         }
                     }
                 } else {
-                    let mut out_frames = Vec::new();
-                    let mut out_len = 0;
+                    // Partition and probe the tables.
+                    for p in partition_idxs.iter_mut() {
+                        p.clear();
+                    }
+                    hash_keys.gen_idxs_per_partition(
+                        &partitioner,
+                        &mut partition_idxs,
+                        &mut [],
+                        emit_unmatched,
+                    );
+
                     for (p, idxs_in_p) in partitions.iter().zip(&partition_idxs) {
                         let mut offset = 0;
                         while offset < idxs_in_p.len() {
+                            let matches_before_limit = probe_limit - probe_match.len() as IdxSize;
                             table_match.clear();
-                            probe_match.clear();
                             offset += p.hash_table.probe_subset(
                                 &hash_keys,
                                 &idxs_in_p[offset..],
@@ -771,66 +935,69 @@ impl ProbeState {
                                 &mut probe_match,
                                 mark_matches,
                                 emit_unmatched,
-                                probe_limit - out_len,
+                                matches_before_limit,
                             ) as usize;
 
                             if table_match.is_empty() {
                                 continue;
                             }
+                            total_matches += table_match.len();
 
-                            // Gather output and send.
-                            let mut build_df = if emit_unmatched {
-                                p.payload.take_opt_chunked_unchecked(&table_match, false)
+                            if emit_unmatched {
+                                build_out.opt_gather_extend(
+                                    &p.payload,
+                                    &table_match,
+                                    ShareStrategy::Always,
+                                );
                             } else {
-                                p.payload
-                                    .take_chunked_unchecked(&table_match, IsSorted::Not, false)
+                                build_out.gather_extend(
+                                    &p.payload,
+                                    &table_match,
+                                    ShareStrategy::Always,
+                                );
                             };
-                            if !payload_rechunked {
-                                // TODO: can avoid rechunk? We have to rechunk here or else we do it
-                                // multiple times during the gather.
-                                payload.rechunk_mut();
-                                payload_rechunked = true;
-                            }
-                            let mut probe_df =
-                                payload.take_slice_unchecked_impl(&probe_match, false);
 
-                            let out_df = if params.left_is_build.unwrap() {
-                                build_df.hstack_mut_unchecked(probe_df.get_columns());
-                                build_df
-                            } else {
-                                probe_df.hstack_mut_unchecked(build_df.get_columns());
-                                probe_df
-                            };
-                            let out_df = postprocess_join(out_df, params);
-
-                            out_len = out_len
-                                .checked_add(out_df.height().try_into().unwrap())
-                                .unwrap();
-                            out_frames.push(out_df);
-
-                            if out_len >= probe_limit {
-                                out_len = 0;
-                                let df =
-                                    accumulate_dataframes_vertical_unchecked(out_frames.drain(..));
-                                let out_morsel = Morsel::new(df, seq, src_token.clone());
-                                if send.send(out_morsel).await.is_err() {
-                                    break;
+                            if probe_match.len() >= probe_limit as usize {
+                                if !payload_rechunked {
+                                    payload.rechunk_mut();
+                                    payload_rechunked = true;
                                 }
+                                probe_out.gather_extend(
+                                    &payload,
+                                    &probe_match,
+                                    ShareStrategy::Always,
+                                );
+                                probe_match.clear();
+                                let out_morsel = new_morsel(&mut build_out, &mut probe_out);
+                                if send.send(out_morsel).await.is_err() {
+                                    return Ok(max_seq);
+                                }
+                                // We had enough matches to need a mid-partition flush, let's assume there are a lot of
+                                // matches and just do a large reserve.
+                                build_out.reserve(probe_limit as usize + max_match_per_key_est);
                             }
                         }
                     }
 
-                    if out_len > 0 {
-                        let df = accumulate_dataframes_vertical_unchecked(out_frames.drain(..));
-                        let out_morsel = Morsel::new(df, seq, src_token.clone());
+                    if !probe_match.is_empty() {
+                        if !payload_rechunked {
+                            payload.rechunk_mut();
+                        }
+                        probe_out.gather_extend(&payload, &probe_match, ShareStrategy::Always);
+                        probe_match.clear();
+                        let out_morsel = new_morsel(&mut build_out, &mut probe_out);
                         if send.send(out_morsel).await.is_err() {
-                            break;
+                            return Ok(max_seq);
                         }
                     }
                 }
             }
 
             drop(wait_token);
+
+            // Move selectivity estimate a bit towards latest value.
+            selectivity_estimate =
+                0.8 * selectivity_estimate + 0.2 * (total_matches as f64 / df_height as f64);
         }
 
         Ok(max_seq)
@@ -838,66 +1005,67 @@ impl ProbeState {
 
     fn ordered_unmatched(
         &mut self,
-        partitioner: &HashPartitioner,
         params: &EquiJoinParams,
     ) -> DataFrame {
-        let mut out_per_partition = Vec::with_capacity(partitioner.num_partitions());
-        let seq_name = PlSmallStr::from_static("__POLARS_PROBE_PRESERVE_ORDER_SEQ");
-        let idx_name = PlSmallStr::from_static("__POLARS_PROBE_PRESERVE_ORDER_IDX");
+        // TODO: parallelize this operator.
+
+        let build_payload_schema = if params.left_is_build.unwrap() {
+            &params.left_payload_schema
+        } else {
+            &params.right_payload_schema
+        };
+
         let mut unmarked_idxs = Vec::new();
+        let mut linearized_idxs = Vec::new();
+
+        for (p_idx, p) in self.table_per_partition.iter().enumerate_idx() {
+            p.hash_table
+                .unmarked_keys(&mut unmarked_idxs, 0, IdxSize::MAX);
+            linearized_idxs.extend(unmarked_idxs.iter().map(|i| {
+                (unsafe { *p.seq_ids.get_unchecked(*i as usize) }, p_idx, *i)
+            }));
+        }
+        
+        linearized_idxs.sort_by_key(|(seq_id, _, _)| *seq_id);
+
         unsafe {
-            for p in self.table_per_partition.iter() {
-                p.hash_table
-                    .unmarked_keys(&mut unmarked_idxs, 0, IdxSize::MAX);
+            let mut build_out = DataFrameBuilder::new(build_payload_schema.clone());
+            build_out.reserve(linearized_idxs.len());
 
-                // Gather and create full-null counterpart.
-                let mut build_df =
-                    p.payload
-                        .take_chunked_unchecked(&unmarked_idxs, IsSorted::Not, false);
-                let len = build_df.height();
-                let mut out_df = if params.left_is_build.unwrap() {
-                    let probe_df = DataFrame::full_null(&params.right_payload_schema, len);
-                    build_df.hstack_mut_unchecked(probe_df.get_columns());
-                    build_df
-                } else {
-                    let mut probe_df = DataFrame::full_null(&params.left_payload_schema, len);
-                    probe_df.hstack_mut_unchecked(build_df.get_columns());
-                    probe_df
-                };
+            // Group indices from the same partition.
+            let mut group_start = 0;
+            let mut gather_idxs = Vec::new();
+            while group_start < linearized_idxs.len() {
+                gather_idxs.clear();
 
-                // The indices are not ordered globally, but within each chunk they are, so sorting
-                // by chunk sequence id, breaking ties by inner chunk idx works.
-                let (chunk_seqs, idx_in_chunk) = unmarked_idxs
-                    .iter()
-                    .map(|chunk_id| {
-                        let (chunk, idx_in_chunk) = chunk_id.extract();
-                        (p.chunk_seq_ids[chunk as usize].to_u64(), idx_in_chunk)
-                    })
-                    .unzip();
+                let (_seq, p_idx, idx_in_p) = linearized_idxs[group_start];
+                gather_idxs.push(idx_in_p);
+                let mut group_end = group_start + 1;
+                while group_end < linearized_idxs.len() && linearized_idxs[group_end].1 == p_idx {
+                    gather_idxs.push(linearized_idxs[group_end].2);
+                    group_end += 1;
+                }
+                
+                build_out.gather_extend(
+                    &self.table_per_partition[p_idx as usize].payload,
+                    &gather_idxs,
+                    ShareStrategy::Never, // Don't keep entire table alive for unmatched indices.
+                );
 
-                let chunk_seqs_ca = UInt64Chunked::from_vec(seq_name.clone(), chunk_seqs);
-                let idxs_ca = IdxCa::from_vec(idx_name.clone(), idx_in_chunk);
-                out_df.with_column_unchecked(chunk_seqs_ca.into_column());
-                out_df.with_column_unchecked(idxs_ca.into_column());
-                out_per_partition.push(out_df);
+                group_start = group_end;
             }
 
-            // Sort by chunk sequence id, then by inner chunk idx.
-            let sort_options = SortMultipleOptions {
-                descending: vec![false],
-                nulls_last: vec![false],
-                multithreaded: true,
-                maintain_order: false,
-                limit: None,
+            let mut build_df = build_out.freeze();
+            let out_df = if params.left_is_build.unwrap() {
+                let probe_df = DataFrame::full_null(&params.right_payload_schema, build_df.height());
+                build_df.hstack_mut_unchecked(probe_df.get_columns());
+                build_df
+            } else {
+                let mut probe_df = DataFrame::full_null(&params.left_payload_schema, build_df.height());
+                probe_df.hstack_mut_unchecked(build_df.get_columns());
+                probe_df
             };
-            let mut out_df = accumulate_dataframes_vertical_unchecked(out_per_partition);
-            out_df
-                .sort_in_place([seq_name.clone(), idx_name.clone()], sort_options)
-                .unwrap();
-            out_df.drop_in_place(&seq_name).unwrap();
-            out_df.drop_in_place(&idx_name).unwrap();
-            out_df = postprocess_join(out_df, params);
-            out_df
+            postprocess_join(out_df, params)
         }
     }
 }
@@ -951,9 +1119,7 @@ impl EmitUnmatchedState {
 
                 // Gather and create full-null counterpart.
                 let out_df = unsafe {
-                    let mut build_df =
-                        p.payload
-                            .take_chunked_unchecked(&unmarked_idxs, IsSorted::Not, false);
+                    let mut build_df = p.payload.take_slice_unchecked_impl(&unmarked_idxs, false);
                     let len = build_df.height();
                     if params.left_is_build.unwrap() {
                         let probe_df = DataFrame::full_null(&params.right_payload_schema, len);
@@ -1038,7 +1204,7 @@ pub struct EquiJoinNode {
     state: EquiJoinState,
     params: EquiJoinParams,
     num_pipelines: usize,
-    table: Option<Box<dyn ChunkedIdxTable>>,
+    table: Option<Box<dyn IdxTable>>,
 }
 
 impl EquiJoinNode {
@@ -1065,9 +1231,9 @@ impl EquiJoinNode {
 
         let table = left_is_build.map(|lib| {
             if lib {
-                new_chunked_idx_table(left_key_schema.clone())
+                new_idx_table(left_key_schema.clone())
             } else {
-                new_chunked_idx_table(right_key_schema.clone())
+                new_idx_table(right_key_schema.clone())
             }
         });
 
@@ -1163,9 +1329,12 @@ impl ComputeNode for EquiJoinNode {
         // If we are building and the build input is done, transition to probing.
         if let EquiJoinState::Build(build_state) = &mut self.state {
             if recv[build_idx] == PortState::Done {
-                self.state = EquiJoinState::Probe(
-                    build_state.finalize(&self.params, self.table.as_deref().unwrap()),
-                );
+                let probe_state = if self.params.preserve_order_build {
+                    build_state.finalize_ordered(&self.params, self.table.as_deref().unwrap())
+                } else {
+                    build_state.finalize_unordered(&self.params, self.table.as_deref().unwrap())
+                };
+                self.state = EquiJoinState::Probe(probe_state);
             }
         }
 
@@ -1176,8 +1345,7 @@ impl ComputeNode for EquiJoinNode {
             if samples_consumed && recv[probe_idx] == PortState::Done {
                 if self.params.emit_unmatched_build() {
                     if self.params.preserve_order_build {
-                        let partitioner = HashPartitioner::new(self.num_pipelines, 0);
-                        let unmatched = probe_state.ordered_unmatched(&partitioner, &self.params);
+                        let unmatched = probe_state.ordered_unmatched(&self.params);
                         let mut src = InMemorySourceNode::new(
                             Arc::new(unmatched),
                             probe_state.max_seq_sent.successor(),
@@ -1337,16 +1505,15 @@ impl ComputeNode for EquiJoinNode {
                 let receivers = recv_ports[build_idx].take().unwrap().parallel();
 
                 build_state
-                    .partitions_per_worker
-                    .resize_with(self.num_pipelines, Vec::new);
+                    .local_builders
+                    .resize_with(self.num_pipelines, Default::default);
                 let partitioner = HashPartitioner::new(self.num_pipelines, 0);
-                for (worker_ps, recv) in build_state.partitions_per_worker.iter_mut().zip(receivers)
-                {
+                for (local_builder, recv) in build_state.local_builders.iter_mut().zip(receivers) {
                     join_handles.push(scope.spawn_task(
                         TaskPriority::High,
                         BuildState::partition_and_sink(
                             recv,
-                            worker_ps,
+                            local_builder,
                             partitioner.clone(),
                             &self.params,
                             state,
