+++ crates/polars-stream/src/nodes/joins/mod.rs
@@ -1,3 +1,2 @@
-pub mod equi_join;
 pub mod in_memory;
-pub mod new_equi_join;
+pub mod equi_join;
