+++ crates/polars-stream/src/physical_plan/to_graph.rs
@@ -809,13 +809,8 @@ fn to_graph_rec<'a>(
                 .map(|e| create_stream_expr(e, ctx, &right_input_schema))
                 .try_collect_vec()?;
 
-            // TODO: implement build-side order-maintaining join in new join impl.
-            let preserve_order_build = matches!(
-                args.maintain_order,
-                MaintainOrderJoin::LeftRight | MaintainOrderJoin::RightLeft
-            );
             ctx.graph.add_node(
-                nodes::joins::new_equi_join::EquiJoinNode::new(
+                nodes::joins::equi_join::EquiJoinNode::new(
                     left_input_schema,
                     right_input_schema,
                     left_key_schema,
